+++ Assets/Resources/Characters/Player/scripts/global/StatsComponent.cs
@@ -6,13 +6,13 @@ public class StatsComponent : MonoBehaviour
     public int maxHp;
     public int str;
     public int agi;
-    public int spd;
+    public float spd;
     public int def;
     public int mgcDef;
     public int atkSpeed;
     public int value;
 
-    public void Initialize(int maxHp, int str, int agi, int spd, int def, int mgcDef, int atkSpeed)
+    public void Initialize(int maxHp, int str, int agi, float spd, int def, int mgcDef, int atkSpeed)
     {
         this.maxHp = maxHp;
         this.str = str;
